Minor refactors (discussed on Discord)
