+++ common/src/main/java/me/shedaniel/architectury/registry/trade/TradeOfferContext.java
@@ -17,17 +17,14 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package me.shedaniel.architectury.registry.trade.interal;
+package me.shedaniel.architectury.registry.trade;
 
-import me.shedaniel.architectury.registry.trade.MerchantOfferAccess;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.item.trading.MerchantOffer;
-import org.jetbrains.annotations.ApiStatus;
 
 import java.util.Random;
 
-@ApiStatus.Internal
-public class TradeOfferContext {
+public abstract class TradeOfferContext {
     private final MerchantOfferAccess offer;
     private final Entity entity;
     private final Random random;
