+++ common/src/main/java/me/shedaniel/architectury/registry/trade/TradeRegistry.java
@@ -20,13 +20,13 @@
 package me.shedaniel.architectury.registry.trade;
 
 import dev.architectury.injectables.annotations.ExpectPlatform;
-import me.shedaniel.architectury.registry.trade.interal.TradeRegistryData;
-import me.shedaniel.architectury.registry.trade.interal.VillagerTradeOfferContext;
-import me.shedaniel.architectury.registry.trade.interal.WanderingTraderOfferContext;
+import me.shedaniel.architectury.registry.trade.impl.TradeRegistryData;
 import net.minecraft.world.entity.npc.VillagerProfession;
 import net.minecraft.world.entity.npc.VillagerTrades;
 
-import java.util.*;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
 import java.util.function.Consumer;
 import java.util.function.Predicate;
 
@@ -56,6 +56,7 @@ public class TradeRegistry {
     
     /**
      * Override the max possible offers a villager can have by its profession and level.
+     *
      * @param profession The Profession of the villager.
      * @param level      The level of the villager. Vanilla range is 1 to 5, however mods may extend that upper limit further.
      * @param maxOffers  Max possible offers a villager can have.
@@ -64,7 +65,7 @@ public class TradeRegistry {
         if (level < 1) {
             throw new IllegalArgumentException("Villager Trade level has to be at least 1!");
         }
-    
+        
         if (maxOffers < 0) {
             throw new IllegalArgumentException("Villager's max offers has to be at least 0!");
         }
@@ -77,6 +78,7 @@ public class TradeRegistry {
     /**
      * Register a callback which provide {@link VillagerTradeOfferContext} to modify the given offer from a villager.
      * The callback gets called when {@link net.minecraft.world.entity.npc.Villager} generates their offer list.
+     *
      * @param callback The callback to handle modification for the given offer context.
      */
     public static void modifyVillagerOffers(Consumer<VillagerTradeOfferContext> callback) {
@@ -87,6 +89,7 @@ public class TradeRegistry {
     /**
      * Register a filter which provide {@link VillagerTradeOfferContext} to test the given offer from a villager.
      * The filter gets called when {@link net.minecraft.world.entity.npc.Villager} generates their offer list.
+     *
      * @param filter The filter to test if an offer should be removed. Returning true means the offer will be removed.
      */
     public static void removeVillagerOffers(Predicate<VillagerTradeOfferContext> filter) {
@@ -97,6 +100,7 @@ public class TradeRegistry {
     /**
      * Register a callback which provide {@link WanderingTraderOfferContext} to modify the given offer from the wandering trader.
      * The callback gets called when {@link net.minecraft.world.entity.npc.WanderingTrader} generates their offer list.
+     *
      * @param callback The callback to handle modification for the given offer context.
      */
     public static void modifyWanderingTraderOffers(Consumer<WanderingTraderOfferContext> callback) {
@@ -107,6 +111,7 @@ public class TradeRegistry {
     /**
      * Register a filter which provide {@link WanderingTraderOfferContext} to test the given offer from the wandering trader.
      * The filter gets called when {@link net.minecraft.world.entity.npc.WanderingTrader} generates their offer list.
+     *
      * @param filter The filter to test if an offer should be removed. Returning true means the offer will be removed.
      */
     public static void removeWanderingTraderOffers(Predicate<WanderingTraderOfferContext> filter) {
@@ -128,13 +133,14 @@ public class TradeRegistry {
     
     /**
      * Override the max possible offers the wandering trader can have. This does not affect the rare trade.
-     * @param maxOffers  Max possible offers a villager can have.
+     *
+     * @param maxOffers Max possible offers a villager can have.
      */
     public static void setWanderingTraderMaxOffers(int maxOffers) {
         if (maxOffers < 0) {
             throw new IllegalArgumentException("Wandering trader's max offers has to be at least 0!");
         }
-    
+        
         TradeRegistryData.wanderingTraderMaxOfferOverride = maxOffers;
     }
 }
