+++ common/src/main/java/me/shedaniel/architectury/registry/trade/WanderingTraderOfferContext.java
@@ -1,4 +1,4 @@
-package me.shedaniel.architectury.registry.trade.interal;
+package me.shedaniel.architectury.registry.trade;
 
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.item.trading.MerchantOffer;
@@ -6,10 +6,10 @@ import org.jetbrains.annotations.ApiStatus;
 
 import java.util.Random;
 
-@ApiStatus.Internal
 public class WanderingTraderOfferContext extends TradeOfferContext {
     private final boolean rare;
     
+    @ApiStatus.Internal
     public WanderingTraderOfferContext(MerchantOffer offer, boolean rare, Entity entity, Random random) {
         super(offer, entity, random);
         this.rare = rare;
